chore: refactor <Logout/>
